make zanata pulling useful
Note that while python-zanata-client *does* accept a comma separated list of
language codes to pull, we ask for 'lang' here as a string, not a list.
This is because publican expects each set of POs to live in a language-specific
directory, so we must pass the --langdir option when pulling strings - one
per language.