first hack at translation building
This loads a builder for every language of every guide.  That's a lot of
builders.  It seems like there might be a less cluttery way to do it,
but we also don't want an isolated failure to affect other things.

There are schedulers too.  One scheduler for each release and language,
and it triggers that language's builders for the given release...
sometime in the middle of the night.

Everything that should build, does. At this point, everything that
should not build (unbranched guides, for example) is spewing failures
into the logs.  Something should be done about that - but it should
happen at runtime, not when the config is loaded, so the helper usage
will have to be creative...