add FedoraHelpers.get_remote_branches()
This commit adds a function to grab remote heads from a given repo. It
then matches the repo names against valid Fedora branch names, and
returns a list of the repo's branches that match.

Using this function to set up schedulers and builders for books will
ensure that only *valid* branches will be monitored and built.