Only read configs directly from sources via cli
If we load an existing metadata file, then merge it with parsed
metadata, the winner might be wrong.  Also, nobody wants to maintain a
separate metadata file and risk it getting out of sync with the info in
the primary content.

As such, this commit removes support for reading pre-existing metadata
(now JSON) during cli-envoked extraction.