Containerized!
Created a short script, rebuild_containers.sh, to help set up a
container environment for development.  It will make sure docker is
installed and running, create a container for the master and run it,
then create a container for the slave and run it.

The master container will be rebuilt if there are any uncommitted
changes to the files used to build it. To force a one time rebuild, as
in for the first build after clone, touch rebuild` in the working directory.