more metadata from docbook
anerist will now extract 'taxonomy' and 'tags' metadata from docbook
sources.  This requires a somewhat specific method for presenting the
metadata:

        <articleinfo>
                <subjectset>
                        <subject>
                                <subjectterm>Scope/Topic</subjectterm>
                        </subject>
                </subjectset>
                <keywordset>
                        <keyword>some word</keyword>
                        <keyword>appropriate</keyword>
                </keywordset>
                ....

All values of `keyword` will be added to the document's metadata.  Only
the first value of subjectterm will be applied, because there is no
implied ordering in the stack, so the heirarchy will be done at the end
of the process to minimize disruption of the taxonomic heirarchy.