lazy add entity values
This would be better with some DTD detection. With that, we could add ie
'/usr/share/sgml/docbook/xml-dtd-4.5/ent/isonum.ent' to the
entity_filelist.

I haven't done that because entities there look like this:

<!ENTITY nbsp             "&#x000A0;" >

and so the code ass is will gleefully substitute in 'u00a0' for '&nbsp;'
because we haven't defined "#x000A0" as an entity either, and I don't
know where that magic comes from.